requirements.txt
aiogram==3.13.1
aiohttp==3.10.5
fastapi==0.115.0
uvicorn==0.31.0
pydantic==2.9.2
python-dotenv==1.0.1
slowapi==0.1.9
redis==5.1.1
rosreestr-api
requests~=2.31.0
httpx~=0.28.1
num2words~=0.5.14